refactor: Move generic coordination auditing out of runtime task locks [ORB-10808]
Planned-By: grok
